% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.mcmc.mvnma.R
\name{as.mcmc.mvnma}
\alias{as.mcmc.mvnma}
\alias{as.mcmc}
\title{Auxiliary function to extract the Markov Chain Monte Carlo object}
\usage{
\method{as.mcmc}{mvnma}(x, drop.reference.group = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{mvnma}.}

\item{drop.reference.group}{A logical indicating whether to drop the MCMC
samples for the reference group.}

\item{\dots}{Additional arguments passed on to
\code{\link[coda]{as.mcmc.list}}.}
}
\value{
The function returns the Markov Chain Monte Carlo object which is a list
of 'mcmc' objects.
}
\description{
Extract the Markov Chain Monte Carlo object from an R object of class
\code{mvnma}.
}
\examples{
# Locate file "mvnma_examples.rda" with mvnma() results
.fname <- system.file("extdata/mvnma_examples.rda", package = "mvnma")
load(.fname)

mcmc12 <- as.mcmc(mvnma12)

library(coda)
oldpar <- par(mfrow = c(3, 3))
# Traceplot
traceplot(mcmc12)
# Density plot
densplot(mcmc12)
# Traceplot and density plot together
plot(mcmc12)

# Do not print the trace plot for the reference group, i.e., placebo
mcmc12.drop <- as.mcmc(mvnma12, drop.reference.group = TRUE)
par(mfrow = c(2, 4))
# traceplot
traceplot(mcmc12.drop)
# density plot
densplot(mcmc12.drop)
# traceplot and density plot together
plot(mcmc12.drop)

par(oldpar)

}
\seealso{
\code{\link{mvnma}}
}
\author{
Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
