% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.mvnma.R
\name{forest.mvnma}
\alias{forest.mvnma}
\title{Forest plot for multivariate network meta-analysis results}
\usage{
\method{forest}{mvnma}(
  x,
  backtransf = FALSE,
  separate = n_distinct(x$sm) > 1,
  leftcols = "studlab",
  leftlabs,
  rightcols = c("effect", "ci"),
  rightlabs,
  col.study = "black",
  col.square = "black",
  col.square.lines = "black",
  col.subgroup = "black",
  squaresize = 0.7,
  header.line = TRUE,
  text.outcome = "Outcome: ",
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link{mvnma}}.}

\item{backtransf}{A logical indicating whether results should be
back transformed in forest plots. If \code{backtransf = TRUE},
results for \code{sm = "OR"} are presented as odds ratios rather
than log odds ratios, for example.}

\item{separate}{A logical indicating whether separate forest plots should be
created for outcomes.}

\item{leftcols}{A character vector specifying (additional) columns to be
printed on the left side of the forest plot or a logical value.}

\item{leftlabs}{A character vector specifying labels for (additional)
columns on left side of the forest plot.}

\item{rightcols}{A character vector specifying (additional) columns to be 
printed on the right side of the forest plot or a logical value.}

\item{rightlabs}{A character vector specifying labels for (additional)
columns on right side of the forest plot.}

\item{col.study}{The colour for treatment effect estimates and confidence
interval limits.}

\item{col.square}{The colour for squares reflecting the treatment effect
estimates.}

\item{col.square.lines}{The colour for the outer lines of squares reflecting
the treatment effect estimates.}

\item{col.subgroup}{The colour to print information on different outcomes.}

\item{squaresize}{The size of squares reflecting the treatment effect
estimates (default: 0.7).}

\item{header.line}{A logical value indicating whether to print a header
line (default: TRUE) or a character string ("both", "below", "").}

\item{text.outcome}{A text string printed before the name of the outcome.}

\item{\dots}{Additional arguments passed on to
\code{\link[meta]{forest.meta}}.}
}
\value{
A forest plot is shown in the active graphics window.
}
\description{
Draws a forest plot in the active graphics window (using grid graphics
system).
}
\examples{
# Locate file "mvnma_examples.rda" with mvnma() results
.fname <- system.file("extdata/mvnma_examples.rda", package = "mvnma")
load(.fname)

# Print multivariate network meta-analysis results
mvnma12

# Generate a forest plot with the results
forest(mvnma12)

# Generate a separate forest plot for each outcome
forest(mvnma12, separate = TRUE)

}
\keyword{hplot}
