% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasse.R
\name{hasse.mvrank}
\alias{hasse.mvrank}
\title{Hasse diagram}
\usage{
\method{hasse}{mvrank}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{mvrank}}.}

\item{\dots}{Additional arguments passed on to
\code{\link[netmeta]{hasse.netposet}}.}
}
\value{
A Hasse diagram is shown in the active graphics window.
}
\description{
This function generates a Hasse diagram for a partial order of treatment
ranks in a multivariate network meta-analysis.
}
\details{
Generate a Hasse diagram (Carlsen & Bruggemann, 2014) for a partial order of
treatment ranks in a network meta-analysis (Rücker & Schwarzer, 2017).

This R function is a wrapper function for
\code{\link[netmeta]{hasse.netposet}}.
}
\examples{
# Locate file "mvnma_examples.rda" with mvnma() results
.fname <- system.file("extdata/mvnma_examples.rda", package = "mvnma")
load(.fname)

# Print odds ratios for efficacy outcomes
outc <- mvnma12$outcomes
#
for (i in outc) {
  cat(paste0("\nOutcome: ", i, "\n\n"))
  print(round(exp(mvnma12[[i]]$TE.random), 2))
}

# Rank treatments using SUCRAs
ranks12 <- mvrank(mvnma12, method = "sucra",
  small.values = c("undes", "undes"))
ranks12

# Get the Hasse diagram for the efficacy outcomes
hasse(ranks12)

}
\references{
Carlsen L, Bruggemann R (2014):
Partial order methodology: a valuable tool in chemometrics.
\emph{Journal of Chemometrics},
\bold{28}, 226--34

Rücker G, Schwarzer G (2017):
Resolve conflicting rankings of outcomes in network meta-analysis:
Partial ordering of treatments.
\emph{Research Synthesis Methods},
\bold{8}, 526--36
}
