% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mvrank.R
\name{plot.mvrank}
\alias{plot.mvrank}
\title{Scatter plot to visualize the ranking of two outcomes in a multivariate network meta-analysis}
\usage{
\method{plot}{mvrank}(
  x,
  which = 1:2,
  pos = 1,
  cex.point = 1,
  cex.label = 0.7,
  pch = 19,
  xlim = c(0, 1),
  ylim = c(0, 1),
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link{mvrank}}.}

\item{which}{A mandatory numeric vector of length 2 specifying which
outcomes should be plotted. For example, setting "outcome = c(2, 3)"
implies that a scatter plot will be generated plotting the rankings
of outcomes 2 and 3.}

\item{pos}{Position of treatment labels.}

\item{cex.point}{a numeric value specifying the size of the points
(default: 1)}

\item{cex.label}{a numeric value specifying the size of the point labels in
the plot (default: 0.7)}

\item{pch}{a vector of plotting characters or symbols (default: 19)}

\item{xlim}{the x limits of the plot}

\item{ylim}{the y limits of the plot}

\item{\dots}{Additional arguments for \code{\link{plot}} function.}
}
\value{
A scatter plot is shown in the active graphics window.
}
\description{
Draw a scatter plot in the active graphics window.
}
\examples{
# Locate file "mvnma_examples.rda" with mvnma() results
.fname <- system.file("extdata/mvnma_examples.rda", package = "mvnma")
load(.fname)

# Rank treatments using SUCRAs (default)
ranks12 <- mvrank(mvnma12, small.values = c("undes", "undes"))
ranks12

# Visualize SUCRAs in a scatter plot with outcome 1 on the x-axis and
# outcome 2 on the y-axis
plot(ranks12)

# Visualize SUCRAs in a scatter plot with outcome 2 on the x-axis and
# outcome 1 on the y-axis
plot(ranks12, which = 2:1)

}
\keyword{hplot}
