% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mvrank.R
\name{print.mvrank}
\alias{print.mvrank}
\title{Print results for treatment ranking in multivariate network meta-analysis}
\usage{
\method{print}{mvrank}(x, digits = gs("digits"), ...)
}
\arguments{
\item{x}{An object of class \code{\link{mvrank}}.}

\item{digits}{Minimal number of significant digits, see
\code{print.default}.}

\item{\dots}{Additional arguments (ignored)}
}
\description{
Print results for treatment ranking in multivariate network meta-analysis
}
\examples{
\donttest{
library("netmeta")

# Use 'pairwise' to obtain contrast based data for the first two outcomes
data("Linde2015")
# Early response
p1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(resp1, resp2, resp3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")
# Early remissions
p2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(remi1, remi2, remi3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Perform analysis considering the efficacy outcomes
p12 <- list(p1, p2)

# Use 'mvdata()' to transform the data in suitable JAGS format
data12 <- mvdata(p12)

# Define outcome labels
outcomes <- c("Early Response", "Early Remission")
 
# Fit the model combining only the two efficacy outcomes
set.seed(1909)
mvnma12 <- mvnma(data = data12, 
  reference.group = "Placebo", outclab = outcomes,
  n.iter = 10, n.burnin = 1)

mvrank(mvnma12, small = c("und", "und"))
}


}
