% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.vikor.R
\name{print.vikor}
\alias{print.vikor}
\title{Evaluate the conditions of the VIKOR method and return the set of compromise
solutions.}
\usage{
\method{print}{vikor}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{\link{vikor}}.}

\item{digits}{A numeric specifying the number of digits to print the
ranking matrix Q.}

\item{\dots}{Additional arguments (ignored).}
}
\description{
This function uses the three ranking matrices Q, S and R obtained by the
VIKOR (or fuzzy VIKOR) method and evaluates the conditions C1 and C2 the
identify the set of compromise solutions.
}
\examples{
\donttest{
# Use 'pairwise' to obtain contrast based data for the first two outcomes

# Early response
pw1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(resp1, resp2, resp3), n = list(n1, n2, n3),
  studlab = id, data = Linde2015, sm = "OR")
# Early remissions
pw2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(remi1, remi2, remi3), n = list(n1, n2, n3),
  studlab = id, data = Linde2015, sm = "OR")

# Define outcome labels
outcomes <- c("Early_Response", "Early_Remission")

# Fit the model combining the two efficacy outcomes
set.seed(1909)
mvnma12 <- mvnma(pw1, pw2,
  reference.group = "Placebo", outclab = outcomes[1:2],
  n.iter = 1000, n.burnin = 100)
mvnma12

# Rank treatments using SUCRAs
ranks12 <- mvrank(mvnma12, method = "sucra",
  small.values = c("undes", "undes"))
ranks12

# Get the best compromise solution across efficacy outcomes
vikor(ranks12)

# Use larger weight for response than remission
vikor(ranks12, weights = c(0.6, 0.3))
}

}
