% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiechart.R
\name{spiechart.mvrank}
\alias{spiechart.mvrank}
\alias{spiechart}
\title{Rank treatments across all outcomes using the spie chart method}
\usage{
\method{spiechart}{mvrank}(x, weights = NULL, ...)

spiechart(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{mvrank}}.}

\item{weights}{Outcome weights. The weights should always sum to 1. If not
then they are standardized. If NULL, the function will assume equal outcome
weights.}

\item{\dots}{Additional arguments passed to \code{\link[ggplot2]{ggplot}}
function.}
}
\value{
A data frame with the area under the spie chart as amalgamated treatment
hierarchy across outcomes.
}
\description{
This function employs the spie chart approach to combine
outcome-specific treatment hierarchies obtained in terms of the
probabilistic ranking metrics:
\itemize{
\item Surface Under the Cumulative Ranking curve (SUCRA), or
\item probability of being best (pbest).
}
}
\details{
The spie chart is a modified pie chart in which each sector corresponds to
a treatment and its radius is proportional to a ranking metric, so that
sector area encodes treatment performance (Daly et al., 2020). This function
constructs one spie chart per outcome using SUCRA or probability of being
best (pbest) metrics, where each metric determines the radius of the
corresponding sector.

An amalgamated treatment hierarchy across outcomes is then derived by
averaging the sector areas across outcomes, providing a quantitative
summary of overall treatment performance.
}
\examples{
# Locate file "mvnma_examples.rda" with mvnma() results
.fname <- system.file("extdata/mvnma_examples.rda", package = "mvnma")
load(.fname)

# Rank treatments using SUCRAs
ranks12 <- mvrank(mvnma12, method = "sucra",
  small.values = c("undes", "undes"))
ranks12

# Get amalgamated treatment hierarchy
spiechart(ranks12)

}
\references{
Daly CH, Mbuagbaw L, Thabane L, Straus SE, Hamid JS (2020):
Spie charts for quantifying treatment effectiveness and safety in multiple
outcome network meta-analysis: a proof-of-concept study.
\emph{BMC Med Res Methodol}, \bold{20}, 266
}
