% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.mcmc.mvnma.R
\name{as.mcmc.mvnma}
\alias{as.mcmc.mvnma}
\alias{as.mcmc}
\title{Auxiliary function to extract the Markov Chain Monte Carlo object}
\usage{
\method{as.mcmc}{mvnma}(x, drop.reference.group = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{mvnma}.}

\item{drop.reference.group}{A logical indicating whether to drop the MCMC
samples for the reference group.}

\item{\dots}{Additional arguments passed on to
\code{\link[coda]{as.mcmc.list}}.}
}
\description{
Extract the Markov Chain Monte Carlo object from an R object of class
\code{mvnma}.
}
\examples{
\donttest{
library("netmeta")

data("Linde2015")

# Use 'pairwise' to obtain contrast based data for each one of the five
# available outcomes 

# Early response
p1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(resp1, resp2, resp3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Early remissions
p2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(remi1, remi2, remi3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Define outcome labels
outcomes <- c("Early_Response", "Early_Remission")
 
# Fit the model combining only the two efficacy outcomes
set.seed(1909)
mvnma12 <- mvnma(p1, p2,
  reference.group = "Placebo", outclab = outcomes,
  n.iter = 1000, n.burnin = 100)
mcmc12 <- as.mcmc(mvnma12)

library(coda)
oldpar <- par(mfrow = c(3, 3))
traceplot(mcmc12)

# Do not print the trace plot for the reference group, i.e., placebo
mcmc12.drop <- as.mcmc(mvnma12, drop.reference.group = TRUE)
par(mfrow = c(2, 4))
traceplot(mcmc12.drop)

par(oldpar)
}

}
\seealso{
\code{\link{mvnma}}
}
\author{
Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
