% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.mvnma.R
\name{forest.mvnma}
\alias{forest.mvnma}
\title{Forest plot for multivariate network meta-analysis results}
\usage{
\method{forest}{mvnma}(
  x,
  backtransf = FALSE,
  leftcols = "studlab",
  leftlabs,
  rightcols = c("effect", "ci"),
  rightlabs,
  col.study = "black",
  col.square = "black",
  col.square.lines = "black",
  squaresize = 0.7,
  header.line = TRUE,
  col.subgroup = "black",
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link{mvnma}}.}

\item{backtransf}{A logical indicating whether results should be back-transformed. 
in forest plots. If backtransf = TRUE, results for sm = "OR".
are presented as odds ratios rather than log odds ratios.}

\item{leftcols}{A character vector specifying (additional) columns to be printed on the left side of the forest plot or a logical value.}

\item{leftlabs}{A character vector specifying labels for (additional) columns on left side of the forest plot.}

\item{rightcols}{A character vector specifying (additional) columns to be printed on the right side of the forest plot or a logical value.}

\item{rightlabs}{A character vector specifying labels for (additional) columns on right side of the forest plot.}

\item{col.study}{The colour for treatment effect estimates and confidence interval limits.}

\item{col.square}{The colour for squares reflecting the treatment effect estimates.}

\item{col.square.lines}{The colour for the outer lines of squares reflecting the treatment effect estimates.}

\item{squaresize}{The size of squares reflecting the treatment effect estimates (default: 0.7).}

\item{header.line}{A logical value indicating whether to print a header line (default: TRUE) or a character string ("both", "below", "").}

\item{col.subgroup}{The colour to print information on different outcomes.}

\item{\dots}{Additional arguments for \code{\link[meta]{forest.meta}} function.}
}
\description{
Draws a forest plot in the active graphics window (using grid graphics
system).
}
\examples{
\donttest{
library("netmeta")

# Use 'pairwise' to obtain contrast based data for the first two outcomes
data("Linde2015")
# Early response
p1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(resp1, resp2, resp3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")
# Early remissions
p2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(remi1, remi2, remi3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Define outcome labels
outcomes <- c("Early Response", "Early Remission")
 
# Fit the model combining the two efficacy outcomes
set.seed(1909)
mvnma12 <- mvnma(p1, p2,
  reference.group = "Placebo", outclab = outcomes,
  n.iter = 1000, n.burnin = 100)
mvnma12

# Generate a forest plot with the results
forest(mvnma12)
}

}
