% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.mvnma.R
\name{forest.mvnma}
\alias{forest.mvnma}
\title{Forest plot for multivariate network meta-analysis}
\usage{
\method{forest}{mvnma}(
  x,
  backtransf = FALSE,
  leftcols = "studlab",
  leftlabs = "Comparison",
  rightcols = c("effect", "ci"),
  rightlabs = c("TE", NA),
  col.study = "black",
  col.square = "black",
  col.square.lines = "black",
  squaresize = 0.7,
  header.line = TRUE,
  col.subgroup = "black",
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link{mvnma}}.}

\item{backtransf}{...}

\item{leftcols}{...}

\item{leftlabs}{...}

\item{rightcols}{...}

\item{rightlabs}{...}

\item{col.study}{...}

\item{col.square}{...}

\item{col.square.lines}{...}

\item{squaresize}{...}

\item{header.line}{...}

\item{col.subgroup}{...}

\item{\dots}{Additional arguments for \code{\link[meta]{forest.meta}} function.}
}
\description{
Draws a forest plot in the active graphics window (using grid graphics system).
}
\examples{
\donttest{
library("netmeta")

# Use 'pairwise' to obtain contrast based data for the first two outcomes
data("Linde2015")
# Early response
p1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(resp1, resp2, resp3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")
# Early remissions
p2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(remi1, remi2, remi3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Perform analysis considering the efficacy outcomes
p12 <- list(p1, p2)

# Use 'mvdata()' to transform the data in suitable JAGS format
data12 <- mvdata(p12)

# Define outcome labels
outcomes <- c("Early Response", "Early Remission")
 
# Fit the model combining only the two efficacy outcomes
set.seed(1909)
mvnma12 <- mvnma(data = data12, 
  reference.group = "Placebo", outclab = outcomes,
  n.iter = 1000, n.burnin = 100)
mvnma12

# Generate a forest plot with the results 
forest(mvnma12)                 
}


}
