% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasse.R
\name{hasse.mvrank}
\alias{hasse.mvrank}
\title{Hasse diagram}
\usage{
\method{hasse}{mvrank}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{mvrank}}.}
}
\description{
This function generates a Hasse diagram for a partial order of treatment ranks in a multivariate network meta-analysis.#'
}
\details{
Generate a Hasse diagram (Carlsen & Bruggemann, 2014) for a partial order of treatment ranks in a network meta-analysis (Rücker & Schwarzer, 2017).

This R function is a wrapper function for a modified version of R function \code{hasse} in R package \bold{hasseDiagram} (Krzysztof Ciomek, \url{https://github.com/kciomek/hasseDiagram}) which is available under the MIT license.
}
\examples{
library(netmeta)

data("Linde2015")

# use 'pairwise' to obtain contrast based data for each one of the five available outcomes 

# Early response

p1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
             event = list(resp1, resp2, resp3), 
              n = list(n1, n2, n3),
              studlab = id,
              data = dat.linde2015,
              sm = "OR")


# Early remissions

p2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
              event = list(remi1, remi2, remi3),
              n = list(n1, n2, n3),
              studlab = id,
              data = dat.linde2015,
              sm = "OR")

# Perform analysis in terms of the Efficacy outcomes

p_effic <- list(p1,p2)

# Use 'mvdata()' to transform the data in suitable JAGS format

data_effic <- mvdata(p_effic)

# Define outcome labels

outlab <- c("Early_Response","Early_Remission")
            
# Fit the model combining only the two efficacy outcomes

mvmodel_effic <- mvnma(data = data_effic,
                reference.group = "Placebo",
                outlab = outlab,
                n.iter = 1000,
                n.burnin = 100)
               

# Rank treatments using sucra

ranks_sucra <- mvrank(mvmodel_effic,small.values = c("undesirable","undesirable"), method = "sucra")
                    
ranks_sucra

# Get the hasse diagram in terms of the Efficacy outcomes

hasse.mvrank(ranks_sucra)
                                                       
}
\references{
Carlsen L, Bruggemann R (2014): Partial order methodology: a valuable tool in chemometrics. \emph{Journal of Chemometrics},\bold{28}, 226--34

Rücker G, Schwarzer G (2017): Resolve conflicting rankings of outcomes in network meta-analysis: Partial ordering of treatments. \emph{Research Synthesis Methods}, \bold{8}, 526--36
}
