% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasse.R
\name{hasse.mvrank}
\alias{hasse.mvrank}
\title{Hasse diagram}
\usage{
\method{hasse}{mvrank}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{mvrank}}.}

\item{\dots}{Additional arguments passed on to
\code{\link[netmeta]{hasse.netposet}}.}
}
\description{
This function generates a Hasse diagram for a partial order of treatment
ranks in a multivariate network meta-analysis.
}
\details{
Generate a Hasse diagram (Carlsen & Bruggemann, 2014) for a partial order of
treatment ranks in a network meta-analysis (Rücker & Schwarzer, 2017).

This R function is a wrapper function for
\code{\link[netmeta]{hasse.netposet}}.
}
\examples{
\donttest{
library("netmeta")

# Use 'pairwise' to obtain contrast based data for the first two outcomes
data("Linde2015")
# Early response
p1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(resp1, resp2, resp3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")
# Early remissions
p2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(remi1, remi2, remi3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Define outcome labels
outcomes <- c("Early Response", "Early Remission")
 
# Fit the model combining the two efficacy outcomes
set.seed(1909)
mvnma12 <- mvnma(p1, p2,
  reference.group = "Placebo", outclab = outcomes,
  n.iter = 1000, n.burnin = 100)
mvnma12

# Print odds ratios for efficacy outcomes
outc <- names(mvnma12)[names(mvnma12) != "cor"]
#
for (i in outc) {
  cat(paste0("\nOutcome: ", i, "\n\n"))
  print(round(exp(mvnma12[[i]]$TE.random), 2))
}

# Rank treatments using SUCRAs
ranks12 <- mvrank(mvnma12, small.values = c("und", "und"), method = "sucra")
ranks12

# Get the Hasse diagram for the efficacy outcomes
hasse(ranks12)
}

}
\references{
Carlsen L, Bruggemann R (2014):
Partial order methodology: a valuable tool in chemometrics.
\emph{Journal of Chemometrics}, \bold{28}, 226--34

Rücker G, Schwarzer G (2017):
Resolve conflicting rankings of outcomes in network meta-analysis:
Partial ordering of treatments.
\emph{Research Synthesis Methods}, \bold{8}, 526--36
}
