% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplot.mvrank.R
\name{heatplot.mvrank}
\alias{heatplot.mvrank}
\title{A heatplot to visualize the output of \code{\link{mvrank}} across all outcomes}
\usage{
\method{heatplot}{mvrank}(
  x,
  sortby = NULL,
  col.num = "white",
  num.size = 4.5,
  col.low = "lightblue",
  col.high = "darkblue",
  width.bar = 1.2,
  angle.x = 0,
  angle.y = 0,
  hjust.x = 0.5,
  hjust.y = 1,
  hjust.legend = 0.5,
  size.x = 8,
  size.y = 10,
  legend.position = "right",
  legend.direction = "vertical",
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link{mvrank}}.}

\item{sortby}{An optional argument to define an outcome to be used as a reference when sorting
the order of treatments on the x-axis. If \code{NULL}, then the first outcome, as it appears in \code{\link{mvrank}},
is used as the reference. If not \code{NULL}, the user can specify this argument using a single numeric value indicating
the position of the outcome in the order of \code{\link{mvrank}} (i.e., setting it equal to 1 will use the first outcome as the reference, as it appears in \code{\link{mvrank}}). Alternatively, the argument can be specified
using a character string explicitly stating the name of the outcome. In this case, the name can be abbreviated.}

\item{col.num}{The color of the numbers in the squares (default: "white").}

\item{num.size}{The size of the numbers in the squares (default: 4.5).}

\item{col.low}{The color for the low end of the gradient (default: "lightblue").}

\item{col.high}{The color for the high end of the gradient (default: "darkblue").}

\item{width.bar}{The width of the bars in the graph (default: 1.2).}

\item{angle.x}{The angle (in [0, 360]) to rotate the text appearing on the x-axis (default: 0).}

\item{angle.y}{The angle (in [0, 360]) to rotate the text appearing on the y-axis (default: 0).}

\item{hjust.x}{Horizontal justification for the text on the x-axis (default: 0.5).}

\item{hjust.y}{Horizontal justification for the text on the y-axis (default: 1).}

\item{hjust.legend}{Horizontal justification for the text in the legend title (default: 0.5).}

\item{size.x}{Text size (in pts) on the x-axis (default: 8).}

\item{size.y}{Text size (in pts) on the y-axis (default: 10).}

\item{legend.position}{Position of the legend. Options are "top", "bottom", "right" (default), and "left". Can be abbreviated.}

\item{legend.direction}{Direction of the legend. Options are "vertical" (default) and "horizontal". Can be abbreviated.}

\item{\dots}{Additional arguments passed to \code{\link[ggplot2]{ggplot}}.}
}
\value{
A \code{ggplot} object.
}
\description{
This function produces a heatplot displaying the results of function \code{\link{mvrank}}. 
The graph can be used to visualize the ranking output when the method used to rank the treatments
is either the SUCRA or the pBV method.
}
\examples{
\donttest{
library("netmeta")

data("Linde2015")

# Use 'pairwise' to obtain contrast based data for each one of the five
# available outcomes 

# Early response
p1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(resp1, resp2, resp3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Early remissions
p2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(remi1, remi2, remi3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Adverse events
p3 <- pairwise(treat = list(treatment1, treatment2,treatment3),
  event = list(ae1, ae2, ae3),  n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Loss to follow-up
p4 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(loss1, loss2, loss3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Loss_to_follow_up_(AE)
p5 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(loss.ae1, loss.ae2, loss.ae3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Define outcome labels
outcomes <- c("Early_Response", "Early_Remission",
  "Adverse_events", "Loss_to_follow_up", "Loss_to_follow_up_AE")

# Fit the model combining all five outcomes
set.seed(1909)
mvnma_all <- mvnma(p1, p2, p3, p4, p5,
  reference.group = "Placebo", outclab = outcomes,
  n.iter = 1000, n.burnin = 100)
# Rank the treatments
  ranks_sucra <- mvrank(mvnma_all, 
  small.values = c("undes", "undes", "des", "des", "des"),
  method = "SUCRA")
ranks_sucra
# Create a heatplot sorting the results according to the first outcome appearing in ranks_sucra (i.e Early_Response)
  heatplot(ranks_sucra,sortby = 1)
# Create a heatplot sorting the results by explicitly mentioning the name of the outcome Early_Response
  heatplot(ranks_sucra,sortby = "Early_Response") 
  }
}
