% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{league}
\alias{league}
\title{Calculate all the outcome specific treatment effect estimates}
\usage{
league(x)
}
\arguments{
\item{x}{An object of class \code{\link{mvnma}}.}
}
\value{
A list of length equal to the total number of outcomes. Each element of the list contains one matrix for all outcome-specific treatment effect estimates and one matrix for all outcome-specific standard deviations of the treatment effect estimates.
}
\description{
This function utilizes the consistency assumption within each outcome. Using the set of basic parameters, it calculates the treatment effect estimates and standard deviations for all outcome-specific treatment comparisons.
}
\examples{
library(netmeta)

data("Linde2015")

# use 'pairwise' to obtain contrast based data for each one of the five available outcomes 

  # Early response

p1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
             event = list(resp1, resp2, resp3), 
              n = list(n1, n2, n3),
              studlab = id,
              data = dat.linde2015,
              sm = "OR")


# Early remissions

p2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
              event = list(remi1, remi2, remi3),
              n = list(n1, n2, n3),
              studlab = id,
              data = dat.linde2015,
              sm = "OR")

# Perform analysis in terms of the Efficacy outcomes

p_effic <- list(p1,p2)

# Use 'mvdata()' to transform the data in suitable JAGS format

data_effic <- mvdata(p_effic)

# Define outcome labels

outlab <- c("Early_Response","Early_Remission")
            
# Fit the model combining only the two efficacy outcomes

mvmodel_effic <- mvnma(data = data_effic,
                reference.group = "Placebo",
                outlab = outlab,
                n.iter = 1000,
                n.burnin = 100)
                
# Extract treatment effect estimates and heterogeneity for Early_Response 

mvmodel_effic$Early_Response$basic_estimates

# Get all estimates                 

league.effic <- league(mvmodel_effic)

# print results in terms of the outcome Early Response 

league.effic$Early_Response

}
