% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnma-package.R
\docType{package}
\name{mvnma-package}
\alias{mvnma-package}
\title{mvnma: Brief overview}
\description{
R package \bold{mvnma} ...
}
\details{
The R package \bold{mvnma} provides the following functions:
\itemize{
\item Function \code{\link{mvnma}} ...
\item Function \code{\link{mvrank}} ...
}

Type \code{help(package = "mvnma")} for a listing of R functions
available in \bold{mvnma}.

Type \code{citation("mvnma")} on how to cite \bold{mvnma}
in publications.

To report problems and bugs, please send an email to Theodoros
Evrenoglou <theodoros.evrenoglou@uniklinik-freiburg.de>.

The development version of \bold{mvnma} is available on GitHub
\url{https://github.com/TEvrenoglou/mvnma}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TEvrenoglou/mvnma}
}

}
\author{
Theodoros Evrenoglou <theodoros.evrenoglou@uniklinik-freiburg.de>,
  Guido Schwarzer <guido.schwarzer@uniklinik-freiburg.de>
}
\keyword{package}
