% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrank.R
\name{mvrank}
\alias{mvrank}
\alias{print.mvrank}
\title{Outcome-specific treatment rankings in multivariate network meta-analysis.}
\usage{
mvrank(x, small.values, method = "SUCRA")

\method{print}{mvrank}(x, digits = gs("digits"), ...)
}
\arguments{
\item{x}{An object of class \code{\link{mvnma}}.}

\item{small.values}{A character vector specifying for each outcome whether
small treatment effects indicate a beneficial ("desirable") or harmful
("undesirable") effect, can be abbreviated.}

\item{method}{The ranking method to be used. Three methods are currently
supported. The SUCRA method (specified as \code{method = "SUCRA"}) is the
default approach. The probability of being best method (specified as
\code{method = "pbest"}) and the mean and median ranks (specified as
\code{method = "ranks"}) are also supported.}

\item{digits}{Minimal number of significant digits, see
\code{print.default}.}

\item{\dots}{Additional arguments (ignored)}
}
\value{
The function returns an object of class \code{mvrank}. It is a list
containing the following components:
\item{ranks}{A named list containing one data frame for each outcome. The
  data frame contains the treatment rankings. For \code{method = "SUCRA"}
  or \code{method = "pbest"}, it contains the variables \code{treatment}
  and the corresponding ranking measure. For \code{method = "ranks"}, it
  contains \code{treatment}, \code{median_rank}, \code{mean_rank},
  \code{lower.CrI}, and \code{upper.CrI}.}
\item{trts}{All treatments occurring in the outcome-specific rankings.}
\item{ranks.shared}{A named list containing rankings recalculated using
  only treatments shared by all outcomes.}
\item{trts.shared}{Treatments occurring in every outcome.}
\item{outcomes}{Outcome labels.}
\item{method}{The ranking method used.}
\item{call, version}{The matched function call and package version used to
  create the object.}
}
\description{
Produces outcome-specific treatment rankings in multivariate
network meta-analysis based on the output of the \code{\link{mvnma}}
function.

This function produces outcome-specific treatment rankings in multivariate
network meta-analysis based on the output of the \code{\link{mvnma}}
function. Two ranking methods (argument \code{method}) are currently
supported (Salanti et al., 2011):
\itemize{
\item Surface under the cumulative ranking curve (SUCRA) method,
\item Probability of being best (pbest) method.
}
}
\examples{
# Locate file "mvnma_examples.rda" with mvnma() results
.fname <- system.file("extdata/mvnma_examples.rda", package = "mvnma")
load(.fname)

# Rank treatments using SUCRAs (default)
ranks_sucra <- mvrank(mvnma_all, 
  small.values = c("undes", "undes", "des", "des", "des"))
#
ranks_sucra

\donttest{
# Rank treatments using pbest
ranks_pbest <- mvrank(mvnma_all,
  small.values = c("undes", "undes", "des", "des", "des"),
  method = "pbest")
#
ranks_pbest         

# Rank treatments using mean and median ranks
ranks_mean_median <- mvrank(mvnma_all,
  small.values = c("undes", "undes", "des", "des", "des"),
  method = "ranks")
#
ranks_mean_median
}

}
\references{
Salanti G, Ades AE, Ioannidis JP (2011):
Graphical methods and numerical summaries for presenting results
from multiple-treatment meta-analysis: an overview and tutorial.
\emph{Journal of Clinical Epidemiology},
\bold{64}, 163--71
}
