% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrank.R
\name{mvrank}
\alias{mvrank}
\title{Outcome-specific treatment rankings in multivariate network meta-analysis.}
\usage{
mvrank(x, small.values, method = "SUCRA")
}
\arguments{
\item{x}{An object of class \code{\link{mvnma}}.}

\item{small.values}{A character vector specifying for each outcome whether
small treatment effects indicate a beneficial ("desirable") or harmful
("undesirable") effect, can be abbreviated.}

\item{method}{The ranking method to be used. Three methods are currently supported.
The SUCRA method (specified as method = "SUCRA") is the default approach. The
probability of best value method (specified as method = "pBV") and the mean and median ranks
(specified as method = "ranks") are also supported.}
}
\description{
This function produces outcome-specific treatment rankings in multivariate
network meta-analysis based on the output of the \code{\link{mvnma}}
function. Two ranking methods are currently supported, these are: (i)
the SUCRA method and (ii) the probability of best value (pBV) method.
}
\examples{
\donttest{
library("netmeta")

data("Linde2015")

# Use 'pairwise' to obtain contrast based data for each one of the five
# available outcomes 

# Early response
p1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(resp1, resp2, resp3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Early remissions
p2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(remi1, remi2, remi3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Adverse events
p3 <- pairwise(treat = list(treatment1, treatment2,treatment3),
  event = list(ae1, ae2, ae3),  n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Loss to follow-up
p4 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(loss1, loss2, loss3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Loss_to_follow_up_(AE)
p5 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(loss.ae1, loss.ae2, loss.ae3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Define outcome labels
outcomes <- c("Early_Response", "Early_Remission",
  "Adverse_events", "Loss_to_follow_up", "Loss_to_follow_up_AE")

# Fit the model combining all five outcomes
set.seed(1909)
mvnma_all <- mvnma(p1, p2, p3, p4, p5,
  reference.group = "Placebo", outclab = outcomes,
  n.iter = 1000, n.burnin = 100)

# Rank treatments using pBV
ranks_pBV <- mvrank(mvnma_all,
  small.values = c("undes", "undes", "des", "des","des"),
  method = "pBV")
ranks_pBV                    
   
# Rank treatments using mean and median ranks
ranks_mean_median <- mvrank(mvnma_all,
  small.values = c("undes", "undes", "des", "des","des"),
  method = "ranks")
ranks_mean_median
   
# Rank treatments using SUCRAs
ranks_sucra <- mvrank(mvnma_all, 
  small.values = c("undes", "undes", "des", "des","des"),
  method = "SUCRA")
ranks_sucra
                     
# Same results without method = "SUCRA" 
ranks_sucra1 <- mvrank(mvnma_all, 
  small.values = c("undes", "undes", "des", "des","des"))
ranks_sucra1
}

}
