% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mvrank.R
\name{plot.mvrank}
\alias{plot.mvrank}
\title{Scatter plot to visualize the ranking lists for two outcomes.}
\usage{
\method{plot}{mvrank}(
  x,
  which = 1:2,
  pos = 1,
  cex.point = 1,
  cex.label = 0.7,
  pch = 19,
  xlim = c(0, 1),
  ylim = c(0, 1),
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link{mvrank}}.}

\item{which}{A mandatory numeric vector of length 2 specifying which
outcomes should be plotted. For example, setting "outcome = c(2,3)"
implies that a scatter plot will be generated plotting the rankings
of outcomes 2 and 3.}

\item{pos}{Position of treatment labels.}

\item{cex.point}{...}

\item{cex.label}{...}

\item{pch}{...}

\item{xlim}{...}

\item{ylim}{...}

\item{...}{Additional arguments for \code{\link{plot}} function.}
}
\description{
Draw a scatter plot (using grid graphics system) in the active graphics
window or store the forest plot in a file.
}
\examples{
\donttest{
library("netmeta")

# Use 'pairwise' to obtain contrast based data for the first two outcomes
data("Linde2015")
# Early response
p1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(resp1, resp2, resp3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")
# Early remissions
p2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(remi1, remi2, remi3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Perform analysis considering the efficacy outcomes
p12 <- list(p1, p2)

# Use 'mvdata()' to transform the data in suitable JAGS format
data12 <- mvdata(p12)

# Define outcome labels
outcomes <- c("Early_Response", "Early_Remission")

# Fit the model combining only the two efficacy outcomes
set.seed(1909)
mvnma12 <- mvnma(data = data12,
  reference.group = "Placebo", outclab = outcomes[1:2],
  n.iter = 1000, n.burnin = 100)
           
# Extract treatment effect estimates and heterogeneity for Early_Response 
mvnma12$Early_Response$basic_estimates
                
# Extract treatment effect estimates and heterogeneity for Early_Response 
mvnma12$Early_Response$basic_estimates

# Get all estimates
league12 <- league(mvnma12)
league12

# Rank treatments using sucra
ranks12 <- mvrank(mvnma12, small.values = c("und","und"), method = "sucra")
ranks12

# Visualize SUCRAs in a scatter plot with outcome 1
# (as specified in the mvdata() function) in the x-axis and outcome 2
# (as specified in the mvdata() function) in the y-axis

plot(ranks12, outcome = c(1, 2))

# Visualize SUCRAs in a scatter plot with outcome 2
# (as specified in the mvdata() function) in the x-axis and outcome 1
# (as specified in the mvdata() function) in the y-axis

plot(ranks12, outcome = c(2, 1))
}

}
