% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mvrank.R
\name{plot.mvrank}
\alias{plot.mvrank}
\title{Scatter plot to visualize the ranking lists in terms of two outcomes.}
\usage{
\method{plot}{mvrank}(x, outcome = NULL, pos = 1, cex.point = 1, cex.label = 0.7, pch = 19, ...)
}
\arguments{
\item{x}{An object of class \code{\link{mvrank}}.}

\item{outcome}{A mandatory numeric vector of length 2 specifying which
outcomes should be plotted. For example, setting "outcome=c(2,3)" implies
that a scatter plot will be generated plotting the rankings of outcomes 2
and 3.}

\item{pos}{...}

\item{cex.point}{...}

\item{cex.label}{...}

\item{pch}{...}

\item{...}{Additional arguments for \code{\link{plot}} function.}
}
\description{
Draw a scatter plot (using grid graphics system) in the active graphics
window or store the forest plot in a file.
}
\examples{
library(netmeta)

data("Linde2015")

# use 'pairwise' to obtain contrast based data for each one of the five available outcomes 

  # Early response

p1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
             event = list(resp1, resp2, resp3), 
              n = list(n1, n2, n3),
              studlab = id,
              data = dat.linde2015,
              sm = "OR")


# Early remissions

p2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
              event = list(remi1, remi2, remi3),
              n = list(n1, n2, n3),
              studlab = id,
              data = dat.linde2015,
              sm = "OR")

# Perform analysis in terms of the Efficacy outcomes

p_effic <- list(p1,p2)

# Use 'mvdata()' to transform the data in suitable JAGS format

data_effic <- mvdata(p_effic)

# Define outcome labels

outlab <- c("Early_Response","Early_Remission")
            
# Fit the model combining only the two efficacy outcomes

mvmodel_effic <- mvnma(data = data_effic,
                reference.group = "Placebo",
                outlab = outlab,
                n.iter = 1000,
                n.burnin = 100)
                
                
                
                
ranks_sucra <- mvrank(mvmodel_effic,small.values = c("undesirable","undesirable"), method = "sucra")
                    
ranks_sucra      

# Visualize SUCRAs in a scatter plot with outcome 1
# (as specified in the mvdata() function) in the x-axis and outcome 2
# (as specified in the mvdata() function) in the y-axis

plot.mvrank(ranks_sucra, outcome = c(1,2))

# Visualize SUCRAs in a scatter plot with outcome 2
# (as specified in the mvdata() function) in the x-axis and outcome 1
# (as specified in the mvdata() function) in the y-axis

plot.mvrank(ranks_sucra, outcome = c(2,1))

}
