% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mvnma.R
\name{print.mvnma}
\alias{print.mvnma}
\title{Print results of multivariate network meta-analysis}
\usage{
\method{print}{mvnma}(
  x,
  digits = gs("digits"),
  digits.sd = gs("digits.sd"),
  print.sd = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link{mvnma}}.}

\item{digits}{Minimal number of significant digits, see
\code{print.default}.}

\item{digits.sd}{Minimal number of significant digits for standard
deviations}

\item{print.sd}{A logical specifying whether standard deviations should be
printed.}

\item{\dots}{Additional arguments (ignored)}
}
\description{
Print results of multivariate network meta-analysis
}
\examples{
\donttest{
library("netmeta")

# Use 'pairwise' to obtain contrast based data for the first two outcomes
data("Linde2015")
# Early response
p1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(resp1, resp2, resp3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")
# Early remissions
p2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(remi1, remi2, remi3), n = list(n1, n2, n3),
  studlab = id, data = dat.linde2015, sm = "OR")

# Define outcome labels
outcomes <- c("Early Response", "Early Remission")
 
# Fit the model combining the two efficacy outcomes
set.seed(1909)
mvnma12 <- mvnma(p1, p2,
  reference.group = "Placebo", outclab = outcomes,
  n.iter = 10, n.burnin = 1)
mvnma12
}

}
