% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiechart.R
\name{spiechart.mvrank}
\alias{spiechart.mvrank}
\alias{spiechart}
\title{Rank treatments across all outcomes using the spie chart method}
\usage{
\method{spiechart}{mvrank}(x, weights = NULL, ...)

spiechart(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{mvrank}}.}

\item{weights}{Outcome weights. The weights should always sum to 1. If not
then they are standardized. If NULL, the function will assume equal outcome
weights.}

\item{\dots}{Additional arguments passed to \code{\link[ggplot2]{ggplot}}
function.}
}
\description{
This function employs the spie chart approach to combine outcome-specific
treatment hierarchies obtained in terms of the probabilistic ranking metrics:
(i) Surface Under the Cumulative Ranking curve (SUCRA) and
(ii) probability of best value (pBV).
}
\details{
The spie chart is a modified pie chart in which each sector corresponds to
a treatment and its radius is proportional to a ranking metric, so that
sector area encodes treatment performance. This function constructs one spie
chart per outcome using SUCRA and probability of best values (pBV) metrics, 
where each metric determines the radius of the corresponding sector. 
An amalgamated treatment hierarchy across outcomes is then derived by
averaging the sector areas across outcomes, providing a quantitative
summary of overall treatment performance.
}
\examples{
\donttest{
# Use 'pairwise' to obtain contrast based data for the first two outcomes
#
data("Linde2015")
# Early response
pw1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(resp1, resp2, resp3), n = list(n1, n2, n3),
  studlab = id, data = Linde2015, sm = "OR")
# Early remissions
pw2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(remi1, remi2, remi3), n = list(n1, n2, n3),
  studlab = id, data = Linde2015, sm = "OR")

# Define outcome labels
outcomes <- c("Early_Response", "Early_Remission")
 
# Fit the model combining only the two efficacy outcomes
set.seed(1909)
mvnma12 <- mvnma(pw1, pw2,
  reference.group = "Placebo", outclab = outcomes,
  n.iter = 1000, n.burnin = 100)
mvnma12

# Rank treatments using SUCRAs
ranks12 <- mvrank(mvnma12, method = "sucra",
  small.values = c("undes", "undes"))
ranks12

# Get amalgamated treatment hierarchy
spiechart(ranks12)
}

}
\references{
Daly CH, Mbuagbaw L, Thabane L, Straus SE, Hamid JS (2020):
Spie charts for quantifying treatment effectiveness and safety in multiple
outcome network meta-analysis: a proof-of-concept study.
\emph{BMC Med Res Methodol}, \bold{20}, 266
}
